# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "LoveAC API",
        "description": "LoveACAPI API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/jwc/fetch_academic_info": {
            "post": {
                "summary": "获取学业信息",
                "description": "获取学术信息（课程数量、绩点等）",
                "operationId": "fetch_academic_info_api_v1_jwc_fetch_academic_info_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthmeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/AcademicInfoResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/provider__loveac__authme__AuthmeResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ],
                                    "title": "Response Fetch Academic Info Api V1 Jwc Fetch Academic Info Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/jwc/fetch_education_plan_info": {
            "post": {
                "summary": "获取培养方案信息",
                "description": "获取培养方案信息",
                "operationId": "fetch_education_plan_info_api_v1_jwc_fetch_education_plan_info_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthmeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/TrainingPlanInfoResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/provider__loveac__authme__AuthmeResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ],
                                    "title": "Response Fetch Education Plan Info Api V1 Jwc Fetch Education Plan Info Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/jwc/fetch_evaluation_course_list": {
            "post": {
                "summary": "获取评教课程列表",
                "description": "获取评教课程列表",
                "operationId": "fetch_evaluation_course_list_api_v1_jwc_fetch_evaluation_course_list_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthmeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/CourseListResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/provider__loveac__authme__AuthmeResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ],
                                    "title": "Response Fetch Evaluation Course List Api V1 Jwc Fetch Evaluation Course List Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/jwc/fetch_exam_info": {
            "post": {
                "summary": "获取考试信息",
                "description": "获取考试信息，包括校统考和其他考试",
                "operationId": "fetch_exam_info_api_v1_jwc_fetch_exam_info_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthmeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ExamInfoAPIResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/provider__loveac__authme__AuthmeResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ],
                                    "title": "Response Fetch Exam Info Api V1 Jwc Fetch Exam Info Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/jwc/evaluation/initialize": {
            "post": {
                "summary": "初始化评价任务",
                "description": "初始化评价任务，获取课程列表",
                "operationId": "initialize_evaluation_task_api_v1_jwc_evaluation_initialize_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthmeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/InitializeResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/provider__loveac__authme__AuthmeResponse"
                                        }
                                    ],
                                    "title": "Response Initialize Evaluation Task Api V1 Jwc Evaluation Initialize Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/jwc/evaluation/start": {
            "post": {
                "summary": "开始评价任务",
                "description": "开始评价任务",
                "operationId": "start_evaluation_task_api_v1_jwc_evaluation_start_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthmeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/TaskOperationResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/provider__loveac__authme__AuthmeResponse"
                                        }
                                    ],
                                    "title": "Response Start Evaluation Task Api V1 Jwc Evaluation Start Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AcademicInfoResponse": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "description": "状态码",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "提示信息",
                        "default": "成功"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/AcademicInfo"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "响应数据"
                    }
                },
                "type": "object",
                "title": "AcademicInfoResponse",
                "description": "学业信息响应"
            },
...
            "AuthmeRequest": {
                "properties": {
                    "token": {
                        "type": "string",
                        "title": "Token"
                    }
                },
                "type": "object",
                "required": [
                    "token"
                ],
                "title": "AuthmeRequest"
            },
...
            "ErrorResponse": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "description": "状态码",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "提示信息",
                        "default": "成功"
                    },
                    "data": {
                        "type": "null",
                        "title": "Data",
                        "description": "响应数据"
                    }
                },
                "type": "object",
                "title": "ErrorResponse",
                "description": "专用错误响应模型"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "provider__loveac__authme__AuthmeResponse": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "title": "AuthmeResponse"
            },
...
        }
    }
}